+++ src/common.f90
@@ -1,9 +1,10 @@
 module xslib_common
 	implicit none
 
-	interface str ! Like python
-		module procedure :: itoa, i8toa, ftoa, f8toa, ctoa, btoa
-	end interface
+	! LEGACY CODE:
+	! interface str
+	! 	module procedure :: itoa, i8toa, ftoa, f8toa, ctoa, btoa
+	! end interface
 
 	interface errorHandling
 		module procedure :: error
@@ -45,9 +46,83 @@ contains
 		return
 	end subroutine warning
 
+	! -------------------------------------------------
+
+	! Trnasforms scalar of any kind to character of shorter possible lenght.
+	recursive function str (value, fmt) result (string)
+		implicit none
+		class(*)					:: value
+		character*(*), optional		:: fmt
+		character*(:), allocatable	:: string
+		character*32				:: deffmt, buffer
+
+		! Act according to the type of variable
+		select type (value)
+		type is (integer)
+			! Default format
+			deffmt = "(i0)"
+			! Overide format if externally defined
+			if (present(fmt)) deffmt = fmt
+			! Write to buffer
+			write (buffer, deffmt) value
+			! Write buffer back to result
+			string = trim(adjustl(buffer))
+
+		type is (integer(KIND=8))
+			deffmt = "(i0)"
+			if (present(fmt)) deffmt = fmt
+			write (buffer, deffmt) value
+			string = trim(adjustl(buffer))
+
+		type is (real)
+			deffmt = "(f0.3)"
+			if (present(fmt)) deffmt = fmt
+			write (buffer, deffmt) value
+			string = trim(adjustl(buffer))
+
+		type is (real(KIND=8))
+			deffmt = "(f0.6)"
+			if (present(fmt)) deffmt = fmt
+			write (buffer, deffmt) value
+			string = trim(adjustl(buffer))
+
+		type is (complex)
+			deffmt = "(f0.3)"
+			if (present(fmt)) deffmt = fmt
+
+			if (aimag(value) > 0) then
+				write (buffer, *) str(real(value), FMT=deffmt)//"+"// &
+				str(aimag(value), FMT=deffmt)//"i"
+			else
+				write (buffer, *) str(real(value), FMT=deffmt)//"-"// &
+				str(abs(aimag(value)), FMT=deffmt)//"i"
+			end if
+			string = trim(adjustl(buffer))
+
+		type is (logical)
+			if (value) then
+				string = "T"
+			else
+				string = "F"
+			end if
+
+		type is (character*(*))
+			string = trim(adjustl(value))
+
+		class default
+			! Can't have everything
+			! buffer = "???"
+			string = "???"
+
+		end select
+
+		! string = trim(adjustl(buffer))
+
+		return
+	end function str
 
 	! -------------------------------------------------
-	! Integer, Real, Complex, logical to string
+	! Individual transformations for Integer, Real, Complex and logical to string
 
 	! Integer to character, in "fmt" format (optional).
 	function itoa (int, fmt) result (string)
@@ -115,9 +190,10 @@ contains
 		return
 	end function f8toa
 
-	function ctoa (imag,fmt) result (res)
+	function ctoa (imag,fmt) result (string)
+		implicit none
 		! Real to character
-		character*(:), allocatable			:: res
+		character*(:), allocatable			:: string
 		complex, intent(in)					:: imag
 		character*(*), intent(in), optional	:: fmt
 		character*32						:: tmp, ifmt
@@ -136,7 +212,7 @@ contains
 			ftoa(abs(aimag(imag)),FMT=ifmt),"i"
 		end if
 
-		res = trim(adjustl(tmp))
+		string = trim(adjustl(tmp))
 
 		return
 	end function ctoa